requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.2
MarkupSafe==1.1.0
python-dotenv==0.10.1
pytz==2018.7
redis==3.0.1
SQLAlchemy==1.2.15
vine==1.1.4
Werkzeug==0.14.1
